kiosque_du_parc/backend/requirements.txt
# KIOSQUE DU PARC - Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-multipart==0.0.12
bcrypt==4.2.0
itsdangerous==2.2.0
jinja2==3.1.4
